functions/requirements.txt
# Azure Functions Python dependencies
# These are installed in the Function App runtime
azure-functions>=1.17.0
azure-eventhub>=5.11.0
azure-storage-file-datalake>=12.14.0
azure-identity>=1.15.0
aiohttp>=3.9.0
httpx>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
structlog>=23.0.0
tenacity>=8.2.0
